add enums.ts file containing all enums; code refactor
